Move wallpaper option out of drawer options to main option page
