+++ app/src/main/java/com/sduduzog/slimlauncher/ui/options/OptionsFragment.kt
@@ -5,6 +5,9 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import android.provider.Settings
+import android.text.SpannableStringBuilder
+import android.text.Spanned
+import android.text.style.TextAppearanceSpan
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -14,6 +17,7 @@ import com.sduduzog.slimlauncher.R
 import com.sduduzog.slimlauncher.ui.dialogs.ChangeThemeDialog
 import com.sduduzog.slimlauncher.ui.dialogs.ChooseTimeFormatDialog
 import com.sduduzog.slimlauncher.utils.BaseFragment
+import com.sduduzog.slimlauncher.utils.isActivityDefaultLauncher
 import kotlinx.android.synthetic.main.options_fragment.*
 
 class OptionsFragment : BaseFragment() {
@@ -56,5 +60,56 @@ class OptionsFragment : BaseFragment() {
         options_fragment_customise_apps.setOnClickListener(Navigation.createNavigateOnClickListener(R.id.action_optionsFragment_to_customiseAppsFragment))
         options_fragment_customize_quick_buttons.setOnClickListener(Navigation.createNavigateOnClickListener(R.id.action_optionsFragment_to_customiseQuickButtonsFragment))
         options_fragment_customize_app_drawer.setOnClickListener(Navigation.createNavigateOnClickListener(R.id.action_optionsFragment_to_customiseAppDrawerFragment))
+    }
+
+    override fun onStart() {
+        super.onStart()
+        // setting up the switch text, since changing the default launcher re-starts the activity
+        // this should able to adapt to it.
+        setupAutomaticDeviceWallpaperSwitch()
+    }
+
+    private fun setupAutomaticDeviceWallpaperSwitch() {
+        val prefsRepo = getUnlauncherDataSource().corePreferencesRepo
+        val appIsDefaultLauncher = isActivityDefaultLauncher(activity)
+        setupDeviceWallpaperSwitchText(appIsDefaultLauncher)
+        options_fragment_auto_device_theme_wallpaper.isEnabled = appIsDefaultLauncher
+
+        prefsRepo.liveData().observe(viewLifecycleOwner) {
+            // always uncheck once app isn't default launcher
+            options_fragment_auto_device_theme_wallpaper.isChecked = appIsDefaultLauncher && it.setThemeWallpaper
+        }
+        options_fragment_auto_device_theme_wallpaper.setOnCheckedChangeListener { _, checked ->
+            prefsRepo.updateSetAutomaticDeviceWallpaper(checked)
+        }
+    }
+
+    /**
+     * Adds a hint text underneath the default text when app is not the default launcher.
+     */
+    private fun setupDeviceWallpaperSwitchText(appIsDefaultLauncher: Boolean) {
+        val text = if (appIsDefaultLauncher) {
+            getText(R.string.customize_app_drawer_fragment_auto_theme_wallpaper_text)
+        } else {
+            buildSwitchTextWithHint()
+        }
+        options_fragment_auto_device_theme_wallpaper.text = text
+    }
+
+    private fun buildSwitchTextWithHint(): CharSequence {
+        val titleText = getText(R.string.customize_app_drawer_fragment_auto_theme_wallpaper_text)
+        // have a title text and a subtitle text to indicate that adapting the
+        // wallpaper can only be done when app it the default launcher
+        val subTitleText = getText(R.string.customize_app_drawer_fragment_auto_theme_wallpaper_subtext_no_default_launcher)
+
+        val spanBuilder = SpannableStringBuilder("$titleText\n$subTitleText")
+        spanBuilder.setSpan(TextAppearanceSpan(context, R.style.TextAppearance_AppCompat_Large), 0, titleText.length, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
+        spanBuilder.setSpan(
+            TextAppearanceSpan(context, R.style.TextAppearance_AppCompat_Small),
+            titleText.length + 1,
+            titleText.length + 1 + subTitleText.length,
+            Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+        )
+        return spanBuilder
     }
 }
