+++ app/src/main/res/layout/options_fragment.xml
@@ -135,6 +135,22 @@
             android:textSize="@dimen/_20ssp"
             app:layout_constraintStart_toStartOf="@+id/options_fragment_customize_quick_buttons"
             app:layout_constraintTop_toBottomOf="@+id/options_fragment_customize_quick_buttons" />
+
+        <androidx.appcompat.widget.SwitchCompat
+            android:id="@+id/options_fragment_auto_device_theme_wallpaper"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/_16sdp"
+            android:layout_marginLeft="@dimen/_16sdp"
+            android:layout_marginTop="32dp"
+            android:layout_marginEnd="@dimen/_16sdp"
+            android:layout_marginRight="@dimen/_16sdp"
+            android:layout_marginBottom="32dp"
+            android:text="@string/customize_app_drawer_fragment_auto_theme_wallpaper_text"
+            android:textAppearance="@style/TextAppearance.AppCompat"
+            android:textSize="@dimen/_18ssp"
+            app:layout_constraintTop_toBottomOf="@id/options_fragment_customize_app_drawer"
+            app:layout_constraintStart_toStartOf="parent"/>
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </ScrollView>
